Proyecto1/backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
boto3==1.34.84
botocore==1.34.84
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
exceptiongroup==1.2.0
fastapi==0.110.1
h11==0.14.0
idna==3.7
jmespath==1.0.1
jwt==1.3.1
psycopg==3.1.18
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
python-dateutil==2.9.0.post0
python-decouple==3.8
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
